Split up io-ts types
